+++ easyshoot/.gitignore
