+++ easyshoot/proguard-rules.pro
